requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Google Gemini
google-genai>=1.0.0

# LlamaIndex workflow engine
llama-index-core>=0.12.0

# Document parsing
docling>=2.0.0
pypdf>=4.0.0
pdfplumber>=0.10.0

# Environment variables
python-dotenv>=1.0.0

# Pydantic (fastapi pulls this in, but pinning for clarity)
pydantic>=2.0.0